fix: Fix some casting issues

+++ src/main/java/by/jackraidenph/dragonsurvival/client/gui/SkinsScreen.java
@@ -455,7 +455,7 @@ public class SkinsScreen extends Screen
 	@Override
 	public boolean mouseScrolled(double mouseX, double mouseY, double amount)
 	{
-		zoom += amount;
+		zoom += (float)amount;
 		zoom = MathHelper.clamp(zoom, 10, 80);
 		
 		return true;
